return type find_distance int32_t

+++ thor.c
@@ -18,7 +18,7 @@
 struct TGPS GPS;
 struct Cords  cords[350]=  {0};
 int cords_index = 0;
-float find_distance(struct Cords c_1, struct Cords c_2){
+int32_t find_distance(struct Cords c_1, struct Cords c_2){
 	return ((c_1.latitude -c_2.latitude)^2 + (c_1.longitude - c_2.longitude)^2)^.5
 }
